BadWord

+++ GestionEvenementBack/src/main/java/tn/eesprit/gestionevenementback/Controllers/LogistiqueController.java
@@ -30,7 +30,7 @@ public class LogistiqueController {
         return logistiqueService.retrieveLogistique(id);
     }
     @GetMapping("/get/event/{id}")
-    Logistique getLogistiqueByEventId(@PathVariable("id") Integer id){
+    Logistique getLogistiqueByEventId(@PathVariable("id") Long id){
         return logistiqueService.retrieveLogistiqueByEventId(id);
     }
     @GetMapping("/all")
