+++ GestionEvenementBack/src/main/java/tn/eesprit/gestionevenementback/Services/ILogistiqueService.java
@@ -8,7 +8,7 @@ public interface ILogistiqueService {
     List<Logistique> retrieveAllLogistiques();
     Logistique addOrUpdateLogistique(Logistique Logistique);
     Logistique retrieveLogistique(Integer id);
-    Logistique retrieveLogistiqueByEventId(Integer id);
+    Logistique retrieveLogistiqueByEventId(Long id);
     Logistique calculDepensesMaterielLogistique(Logistique logistique);
     void removeLogistique(Integer id);
 }
